@@ -1,50 +1,51 @@
-import React, { Component } from "react";
-import Slider from "../src/slider";
+import React, { useState, useRef } from "react";
+import Slider from "react-slick";
 import { baseUrl } from "./config";
 
-export default class SlickGoTo extends React.Component {
-  state = {
-    slideIndex: 0,
-    updateCount: 0
+function SlickGoTo() {
+  const [slideIndex, setSlideIndex] = useState(0);
+  const [updateCount, setUpdateCount] = useState(0);
+  let sliderRef = useRef(null);
+  const settings = {
+    dots: false,
+    infinite: true,
+    speed: 500,
+    slidesToShow: 1,
+    slidesToScroll: 1,
+    afterChange: () => setUpdateCount(updateCount + 1),
+    beforeChange: (current, next) => setSlideIndex(next)
   };
-
-  render() {
-    const settings = {
-      dots: false,
-      infinite: true,
-      speed: 500,
-      slidesToShow: 1,
-      slidesToScroll: 1,
-      afterChange: () =>
-        this.setState(state => ({ updateCount: state.updateCount + 1 })),
-      beforeChange: (current, next) => this.setState({ slideIndex: next })
-    };
-    return (
-      <div>
-        <h2>Slick Go To</h2>
-        <p>Total updates: {this.state.updateCount} </p>
-        <input
-          onChange={e => this.slider.slickGoTo(e.target.value)}
-          value={this.state.slideIndex}
-          type="range"
-          min={0}
-          max={3}
-        />
-        <Slider ref={slider => (this.slider = slider)} {...settings}>
-          <div>
-            <img src={baseUrl + "/abstract01.jpg"} />
-          </div>
-          <div>
-            <img src={baseUrl + "/abstract02.jpg"} />
-          </div>
-          <div>
-            <img src={baseUrl + "/abstract03.jpg"} />
-          </div>
-          <div>
-            <img src={baseUrl + "/abstract04.jpg"} />
-          </div>
-        </Slider>
-      </div>
-    );
-  }
+  return (
+    <div className="slider-container">
+      <p>Total updates: {updateCount} </p>
+      <input
+        onChange={e => sliderRef.slickGoTo(e.target.value)}
+        value={slideIndex}
+        type="range"
+        min={0}
+        max={3}
+      />
+      <Slider
+        ref={slider => {
+          sliderRef = slider;
+        }}
+        {...settings}
+      >
+        <div>
+          <img src={baseUrl + "/abstract01.jpg"} />
+        </div>
+        <div>
+          <img src={baseUrl + "/abstract02.jpg"} />
+        </div>
+        <div>
+          <img src={baseUrl + "/abstract03.jpg"} />
+        </div>
+        <div>
+          <img src={baseUrl + "/abstract04.jpg"} />
+        </div>
+      </Slider>
+    </div>
+  );
 }
+
+export default SlickGoTo;
