@@ -4,7 +4,7 @@
 
 sinon.stub(console, "error");
 
-import { render, shallow } from "enzyme";
+import { render } from "@testing-library/react";
 import React from "react";
 import sinon from "sinon";
 
@@ -22,48 +22,52 @@ function CustomArrow(props) {
 
 describe("Previous arrows", () => {
   it("should render arrow", () => {
-    const wrapper = shallow(<PrevArrow />);
-    expect(wrapper.find("button")).toHaveLength(1);
+    const { container } = render(<PrevArrow />);
+    expect(Array.from(container.getElementsByTagName("button"))).toHaveLength(
+      1
+    );
   });
 
   it("should not result in errors", () => {
-    shallow(<PrevArrow />);
+    render(<PrevArrow />);
 
     expect(console.error.called).toBe(false);
   });
 
-  // it('should pass slide data to custom arrow', () => {
-  //   let elAttributes;
-  //   let arr = <CustomArrow />
-  //
-  //   const wrapper = render(<PrevArrow currentSlide={3} prevArrow={arr} slideCount={5} />);
-  //
-  //   elAttributes = wrapper.find('.sample')[0].attribs;
-  //   expect(elAttributes['data-currentslide']).toBe('3');
-  //   expect(elAttributes['data-slidecount']).toBe('5');
-  // });
+  //  it('should pass slide data to custom arrow', () => {
+  //    let elAttributes;
+  //    let arr = <CustomArrow />
+
+  //    const {container}= render(<PrevArrow currentSlide={3} prevArrow={arr} slideCount={5} />);
+
+  //    elAttributes =x=> container.querySelectorAll('.sample')[0].getAttribute(x);
+  //    expect(elAttributes('data-currentslide')).toBe('3');
+  //    expect(elAttributes('data-slidecount')).toBe('5');
+  //  });
 });
 
 describe("Next arrows", () => {
   it("should render arrow", () => {
-    const wrapper = shallow(<NextArrow />);
-    expect(wrapper.find("button")).toHaveLength(1);
+    const { container } = render(<NextArrow />);
+    expect(Array.from(container.getElementsByTagName("button"))).toHaveLength(
+      1
+    );
   });
 
-  // it('should not result in errors', () => {
-  //   shallow(<NextArrow />);
-  //
-  //   expect(console.error.called).toBe(false);
-  // });
+  //  it('should not result in errors', () => {
+  //    render(<NextArrow />);
 
-  // it('should pass slide data to custom arrow', () => {
-  //   let elAttributes;
-  //   let arr = <CustomArrow />
-  //
-  //   const wrapper = render(<NextArrow currentSlide={6} nextArrow={arr} slideCount={9} />);
-  //
-  //   elAttributes = wrapper.find('.sample')[0].attribs;
-  //   expect(elAttributes['data-currentslide']).toBe('6');
-  //   expect(elAttributes['data-slidecount']).toBe('9');
-  // });
+  //    expect(console.error.called).toBe(false);
+  //  });
+
+  //  it('should pass slide data to custom arrow', () => {
+  //    let elAttributes;
+  //    let arr = <CustomArrow />
+
+  //    const {container} = render(<NextArrow currentSlide={6} nextArrow={arr} slideCount={9} />);
+
+  //    elAttributes =(x)=> container.querySelectorAll('.sample')[0].getAttribute(x);
+  //    expect(elAttributes('data-currentslide')).toBe('6');
+  //    expect(elAttributes('data-slidecount')).toBe('9');
+  //  });
 });
