@@ -1,25 +1,31 @@
+import {
+  clickNext,
+  clickPrevious,
+  getSlides,
+  getCurrentSlide,
+  hasClass,
+  getClonesCount
+} from "../test-utils";
 import { createInnerSliderWrapper } from "./testUtils";
 
 // given slider and actions objects, and simulates given actions
 function simulateActions(slider, actions) {
-  const nextArrow = slider.find(".slick-next");
-  const prevArrow = slider.find(".slick-prev");
   if (actions.clickNext) {
     for (let click = 0; click < actions.clickNext; click++) {
-      nextArrow.simulate("click");
+      clickNext(slider);
     }
   }
   if (actions.clickPrev) {
     for (let click = 0; click < actions.clickPrev; click++) {
-      prevArrow.simulate("click");
+      clickPrevious(slider);
     }
   }
   if (actions.clickSequence) {
     for (let click of actions.clickSequence) {
       if (click === "n") {
-        nextArrow.simulate("click");
+        clickNext(slider);
       } else if (click === "p") {
-        prevArrow.simulate("click");
+        clickPrevious(slider);
       } else {
         // not a valid action for now
       }
@@ -28,10 +34,13 @@ function simulateActions(slider, actions) {
   // console.log('after simulating actions, state of slider:', slider.state())
 }
 
-function fetchDetails(slider, keys) {
-  let details = { ...fetchDOMDetails(slider, keys) };
+function fetchDetails(container, keys) {
+  let details = { ...fetchDOMDetails(container, keys) };
   if (keys.currentSlide) {
-    details.currentSlide = slider.state().currentSlide.toString();
+    // details.currentSlide = getCurrentSlide(container).textContent;
+    details.currentSlide = parseInt(
+      getCurrentSlide(container).getAttribute("data-index")
+    );
   }
   return details;
 }
@@ -43,20 +52,24 @@ function fetchDOMDetails(slider, keys) {
     allSlides = [],
     clonedSlides = [],
     visibleSlides = []; // currently no way to find these
-  slider.find("div.slick-slide").forEach((slide, index) => {
+  //Array.from(getSlides(slider)).forEach(e=>console.log(Array.from(e.getElementsByTagName("h3")).length));
+  Array.from(getSlides(slider)).forEach((slide, index) => {
     const slideObj = {
-      index: slide.prop("data-index").toString(),
-      value: slide.find("h3").length === 1 ? slide.find("h3").text() : "..."
+      index: slide.getAttribute("data-index"),
+      value:
+        Array.from(slide.getElementsByTagName("h3")).length === 1
+          ? slide.querySelector("h3").textContent
+          : "..."
     };
     allSlides.push(slideObj);
-    if (slide.hasClass("slick-active")) {
+    if (hasClass(slide, "slick-current")) {
+      currentSlide = slideObj;
+    }
+    if (hasClass(slide, "slick-active")) {
       activeSlides.push(slideObj);
     }
-    if (slide.hasClass("slick-cloned")) {
+    if (hasClass(slide, "slick-cloned")) {
       clonedSlides.push(slideObj);
-    }
-    if (slide.hasClass("slick-current")) {
-      currentSlide = slideObj;
     }
   });
   if (keys.currentSlide) {
