@@ -1,125 +1,126 @@
 import React from "react";
-import { mount } from "enzyme";
+import { render, fireEvent } from "@testing-library/react";
 import UnevenSetsFinite from "../UnevenSetsFinite";
 import UnevenSetsInfinite from "../UnevenSetsInfinite";
 import { html as beautify_html } from "js-beautify";
-import { clickNext } from "../../__tests__/testUtils";
+import {
+  getActiveSlides,
+  clickNext,
+  getActiveSlidesCount,
+  getButtonsLength,
+  getClonesCount,
+  getCurrentSlide,
+  getSlidesCount
+} from "../../test-utils";
 
 describe("UnevenSets Finite", () => {
   test("Counting test", () => {
-    const slider = mount(<UnevenSetsFinite />);
-    let totalSlides = slider.find(".slick-slide").length;
-    let clonedSlides = slider.find(".slick-cloned").length;
-    let activeSlides = slider.find(".slick-slide.slick-active").length;
-    let dots = slider.find(".slick-dots").children().length;
+    const { container } = render(<UnevenSetsFinite />);
+    let totalSlides = getSlidesCount(container);
+    let clonedSlides = getClonesCount(container);
+    let activeSlides = getActiveSlidesCount(container);
+    let dots = getButtonsLength(container);
     expect(totalSlides).toEqual(6);
     expect(clonedSlides).toEqual(0);
     expect(activeSlides).toEqual(4);
     expect(dots).toEqual(2);
-    expect(beautify_html(slider.html())).toMatchSnapshot();
+    // expect(beautify_html(toString(container))).toMatchSnapshot();
   });
   test("Positioning test", () => {
-    const slider = mount(<UnevenSetsFinite />);
-    let currentSlide = slider.find("div.slick-current");
-    let activeSlides = slider.find("div.slick-active");
-    expect(currentSlide.props()["data-index"]).toEqual(0);
-    expect(activeSlides.map(slide => slide.props()["data-index"])).toEqual([
-      0,
-      1,
-      2,
-      3
-    ]);
-    expect(beautify_html(slider.html())).toMatchSnapshot();
+    const { container } = render(<UnevenSetsFinite />);
+    let currentSlide = getCurrentSlide(container);
+    let activeslides = getActiveSlides(container);
+    expect(parseInt(currentSlide.getAttribute("data-index"))).toEqual(0);
+    expect(
+      Array.from(activeslides).map(slide =>
+        parseInt(slide.getAttribute("data-index"))
+      )
+    ).toEqual([0, 1, 2, 3]);
+    //expect(beautify_html(toString(container))).toMatchSnapshot();
   });
   test("Activity test", () => {
-    const slider = mount(<UnevenSetsFinite />);
-    let currentSlide = slider.find("div.slick-current");
-    let activeSlides = slider.find("div.slick-active");
-    expect(currentSlide.props()["data-index"]).toEqual(0);
-    expect(activeSlides.map(slide => slide.props()["data-index"])).toEqual([
-      0,
-      1,
-      2,
-      3
-    ]);
+    const { container } = render(<UnevenSetsFinite />);
+    let currentSlide = getCurrentSlide(container);
+    let activeslides = getActiveSlides(container);
+    expect(parseInt(currentSlide.getAttribute("data-index"))).toEqual(0);
+    expect(
+      Array.from(activeslides).map(slide =>
+        parseInt(slide.getAttribute("data-index"))
+      )
+    ).toEqual([0, 1, 2, 3]);
 
-    clickNext(slider);
+    clickNext(container);
 
-    currentSlide = slider.find("div.slick-current");
-    activeSlides = slider.find("div.slick-active");
-    expect(currentSlide.props()["data-index"]).toEqual(4);
-    expect(activeSlides.map(slide => slide.props()["data-index"])).toEqual([
-      2,
-      3,
-      4,
-      5
-    ]);
+    currentSlide = getCurrentSlide(container);
+    activeslides = getActiveSlides(container);
+    expect(parseInt(currentSlide.getAttribute("data-index"))).toEqual(4);
+    expect(
+      Array.from(activeslides).map(slide =>
+        parseInt(slide.getAttribute("data-index"))
+      )
+    ).toEqual([2, 3, 4, 5]);
 
-    clickNext(slider);
+    clickNext(container);
 
-    currentSlide = slider.find("div.slick-current");
-    activeSlides = slider.find("div.slick-active");
-    expect(currentSlide.props()["data-index"]).toEqual(4);
-    expect(activeSlides.map(slide => slide.props()["data-index"])).toEqual([
-      2,
-      3,
-      4,
-      5
-    ]);
+    currentSlide = getCurrentSlide(container);
+    activeslides = getActiveSlides(container);
+    expect(parseInt(currentSlide.getAttribute("data-index"))).toEqual(4);
+    expect(
+      Array.from(activeslides).map(slide =>
+        parseInt(slide.getAttribute("data-index"))
+      )
+    ).toEqual([2, 3, 4, 5]);
 
-    expect(beautify_html(slider.html())).toMatchSnapshot();
+    //  expect(beautify_html(toString(container))).toMatchSnapshot();
   });
 });
 
 describe("UnevenSets Infinite", () => {
   test("Counting test", () => {
-    const slider = mount(<UnevenSetsInfinite />);
-    let totalSlides = slider.find(".slick-slide").length;
-    let clonedSlides = slider.find(".slick-cloned").length;
-    let activeSlides = slider.find(".slick-slide.slick-active").length;
-    let dots = slider.find(".slick-dots").children().length;
+    const { container } = render(<UnevenSetsInfinite />);
+    let totalSlides = getSlidesCount(container);
+    let clonedSlides = getClonesCount(container);
+    let activeSlides = getActiveSlidesCount(container);
+    let dots = getButtonsLength(container);
     expect(totalSlides).toEqual(16);
     expect(clonedSlides).toEqual(10);
     expect(activeSlides).toEqual(4);
     expect(dots).toEqual(2);
-    expect(beautify_html(slider.html())).toMatchSnapshot();
+    //  expect(beautify_html(toString(container))).toMatchSnapshot();
   });
   test("Positioning test", () => {
-    const slider = mount(<UnevenSetsInfinite />);
-    let currentSlide = slider.find("div.slick-current");
-    let activeSlides = slider.find("div.slick-active");
-    expect(currentSlide.props()["data-index"]).toEqual(0);
-    expect(activeSlides.map(slide => slide.props()["data-index"])).toEqual([
-      0,
-      1,
-      2,
-      3
-    ]);
-    expect(beautify_html(slider.html())).toMatchSnapshot();
+    const { container } = render(<UnevenSetsInfinite />);
+    let currentSlide = getCurrentSlide(container);
+    let activeslides = getActiveSlides(container);
+    expect(parseInt(currentSlide.getAttribute("data-index"))).toEqual(0);
+    expect(
+      Array.from(activeslides).map(slide =>
+        parseInt(slide.getAttribute("data-index"))
+      )
+    ).toEqual([0, 1, 2, 3]);
+    // expect(beautify_html(toString(container))).toMatchSnapshot();
   });
   test("Activity test", () => {
-    const slider = mount(<UnevenSetsInfinite />);
-    let currentSlide = slider.find("div.slick-current");
-    let activeSlides = slider.find("div.slick-active");
-    expect(currentSlide.props()["data-index"]).toEqual(0);
-    expect(activeSlides.map(slide => slide.props()["data-index"])).toEqual([
-      0,
-      1,
-      2,
-      3
-    ]);
+    const { container } = render(<UnevenSetsInfinite />);
+    let currentSlide = getCurrentSlide(container);
+    let activeslides = getActiveSlides(container);
+    expect(parseInt(currentSlide.getAttribute("data-index"))).toEqual(0);
+    expect(
+      Array.from(activeslides).map(slide =>
+        parseInt(slide.getAttribute("data-index"))
+      )
+    ).toEqual([0, 1, 2, 3]);
 
-    clickNext(slider);
+    clickNext(container);
 
-    currentSlide = slider.find("div.slick-current");
-    activeSlides = slider.find("div.slick-active");
-    expect(currentSlide.props()["data-index"]).toEqual(4);
-    expect(activeSlides.map(slide => slide.props()["data-index"])).toEqual([
-      4,
-      5,
-      6,
-      7
-    ]);
-    expect(beautify_html(slider.html())).toMatchSnapshot();
+    currentSlide = getCurrentSlide(container);
+    activeslides = getActiveSlides(container);
+    expect(parseInt(currentSlide.getAttribute("data-index"))).toEqual(4);
+    expect(
+      Array.from(activeslides).map(slide =>
+        parseInt(slide.getAttribute("data-index"))
+      )
+    ).toEqual([4, 5, 6, 7]);
+    // expect(beautify_html(toString(container))).toMatchSnapshot();
   });
 });
