@@ -1,4 +1,6 @@
-export const baseUrl =
-  process.env.NODE_ENV === "production"
-    ? "https://s3.amazonaws.com/static.neostack.com/img/react-slick"
-    : "/img/react-slick";
+// export const baseUrl =
+//   process.env.NODE_ENV === "production"
+//     ? "https://s3.amazonaws.com/static.neostack.com/img/react-slick"
+//     : "/img/react-slick";
+
+export const baseUrl = "/img/react-slick";
