@@ -50,7 +50,9 @@ let defaultProps = {
   useTransform: true,
   variableWidth: false,
   vertical: false,
-  waitForAnimate: true
+  waitForAnimate: true,
+  asNavFor: null,
+  unslick: false
 };
 
 export default defaultProps;
