@@ -35,6 +35,7 @@ export class InnerSlider extends React.Component {
     this.state = {
       ...initialState,
       currentSlide: this.props.initialSlide,
+      targetSlide: this.props.initialSlide ? this.props.initialSlide : 0,
       slideCount: React.Children.count(this.props.children)
     };
     this.callbackTimers = [];
@@ -133,7 +134,8 @@ export class InnerSlider extends React.Component {
       }
       if (
         typeof prevProps[key] === "object" ||
-        typeof prevProps[key] === "function"
+        typeof prevProps[key] === "function" ||
+        isNaN(prevProps[key])
       ) {
         continue;
       }
@@ -304,7 +306,8 @@ export class InnerSlider extends React.Component {
   };
   checkImagesLoad = () => {
     let images =
-      (this.list && this.list.querySelectorAll &&
+      (this.list &&
+        this.list.querySelectorAll &&
         this.list.querySelectorAll(".slick-slide img")) ||
       [];
     let imagesCount = images.length,
@@ -316,8 +319,8 @@ export class InnerSlider extends React.Component {
         image.onclick = () => image.parentNode.focus();
       } else {
         const prevClickHandler = image.onclick;
-        image.onclick = () => {
-          prevClickHandler();
+        image.onclick = e => {
+          prevClickHandler(e);
           image.parentNode.focus();
         };
       }
